Change leveltype string to an enum and make it easier to access from other classes
